# labels: make InternStrings a no-op for stringlabels version