@cstyan 2024-02-05
LGTM

curious if now the compiler would do the smart thing when this build tag is used, and just drops the intern calls altogether?