@bboreham 2024-02-06
Yes it should inline at the point of call, which will result in no code generated.